Complete bird bullet for bird enemy (have not used yet)

+++ Metroid/BulletObject.h
@@ -35,7 +35,7 @@ public:
 
 	//============ VIRTUAL METHOD =================
 	virtual void InitSprites(LPDIRECT3DDEVICE9 d3ddv);
-	virtual void Update(int t);
+	virtual void Update(float t);
 	virtual void Render();
 	virtual void Shoot(BULLET_DIRECTION dir);
 	virtual void Shoot(BULLET_DIRECTION dir, float posX, float posY);
